Use master branch of chatterbox
Chatterbox in version 0.5.0 has a memory leak which has been there
since at least December 2016. The problem is that finished HTTP/2
streams are not cleared out properly when you get their response. This
means the Chatterbox process will keep allocating more and more
memory as time goes on and eventually crash the Erlang node.

Use the git-version of Chatterbox for now, which has the fix in place.
This avoids memleaks of this kind.